refactor : 모델별로 기능 분리
